Enum pallet_grandpa::pallet::Error

pub enum Error<T> {
    PauseFailed,
    ResumeFailed,
    ChangePending,
    TooSoon,
    InvalidKeyOwnershipProof,
    InvalidEquivocationProof,
    DuplicateOffenceReport,
    // some variants omitted
}

Custom dispatch errors of this pallet.

Variants

PauseFailed

Attempt to signal GRANDPA pause when the authority set isn’t live (either paused or already pending pause).

ResumeFailed

Attempt to signal GRANDPA resume when the authority set isn’t paused (either live or already pending resume).

ChangePending

Attempt to signal GRANDPA change with one already pending.

TooSoon

Cannot signal forced change so soon after last.

InvalidKeyOwnershipProof

A key ownership proof provided as part of an equivocation report is invalid.

InvalidEquivocationProof

An equivocation proof provided as part of an equivocation report is invalid.

DuplicateOffenceReport

A given equivocation report is valid but already previously reported.