Trait pallet_grandpa::pallet::Config

pub trait Config: Config {
    type Event: From<Event> + Into<<Self as Config>::Event> + IsType<<Self as Config>::Event>;
    type Call: From<Call<Self>>;
    type KeyOwnerProof: Parameter + GetSessionNumber + GetValidatorCount;
    type KeyOwnerIdentification: Parameter;
    type KeyOwnerProofSystem: KeyOwnerProofSystem<(KeyTypeId, AuthorityId), Proof = Self::KeyOwnerProof, IdentificationTuple = Self::KeyOwnerIdentification>;
    type HandleEquivocation: HandleEquivocation<Self>;
    type WeightInfo: WeightInfo;
    type MaxAuthorities: Get<u32>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event> + Into<<Self as Config>::Event> + IsType<<Self as Config>::Event>

The event type of this module.

type Call: From<Call<Self>>

The function call.

type KeyOwnerProof: Parameter + GetSessionNumber + GetValidatorCount

The proof of key ownership, used for validating equivocation reports The proof must include the session index and validator count of the session at which the equivocation occurred.

type KeyOwnerIdentification: Parameter

The identification of a key owner, used when reporting equivocations.

type KeyOwnerProofSystem: KeyOwnerProofSystem<(KeyTypeId, AuthorityId), Proof = Self::KeyOwnerProof, IdentificationTuple = Self::KeyOwnerIdentification>

A system for proving ownership of keys, i.e. that a given key was part of a validator set, needed for validating equivocation reports.

type HandleEquivocation: HandleEquivocation<Self>

The equivocation handling subsystem, defines methods to report an offence (after the equivocation has been validated) and for submitting a transaction to report an equivocation (from an offchain context). NOTE: when enabling equivocation handling (i.e. this type isn’t set to ()) you must use this pallet’s ValidateUnsigned in the runtime definition.

type WeightInfo: WeightInfo

Weights for this pallet.

type MaxAuthorities: Get<u32>

Max Authorities in use