pallet_grandpa

Struct EquivocationReportSystem

pub struct EquivocationReportSystem<T, R, P, L>(/* private fields */);

GRANDPA equivocation offence report system.

This type implements OffenceReportSystem such that:

Equivocation reports are published on-chain as unsigned extrinsic via offchain::SendTransactionTypes.

On-chain validity checks and processing are mostly delegated to the user provided generic types implementing KeyOwnerProofSystem and ReportOffence traits.

Offence reporter for unsigned transactions is fetched via the the authorship pallet.