Struct pallet_grandpa::Module

pub struct Module<T: Config>(_);

Implementations

impl<T: Config + 'static> Module<T>

pub fn state() -> StoredState<T::BlockNumber>

State of the current authority set.

pub fn pending_change() -> Option<StoredPendingChange<T::BlockNumber>>

Pending change: (signaled at, scheduled change).

pub fn next_forced() -> Option<T::BlockNumber>

next block number where we can force a change.

pub fn stalled() -> Option<(T::BlockNumber, T::BlockNumber)>

true if we are currently stalled.

pub fn current_set_id() -> SetId

The number of changes (both in terms of keys and underlying economic responsibilities) in the “set” of Grandpa validators from genesis.

pub fn session_for_set<K: EncodeLike<SetId>>(key: K) -> Option<SessionIndex>

A mapping from grandpa set ID to the index of the most recent session for which its members were responsible.

TWOX-NOTE: SetId is not under user control.

impl<T: Config> Module<T>

pub fn grandpa_authorities() -> AuthorityList

Get the current set of authorities, along with their respective weights.

pub fn schedule_pause(in_blocks: T::BlockNumber) -> DispatchResult

Schedule GRANDPA to pause starting in the given number of blocks. Cannot be done when already paused.

pub fn schedule_resume(in_blocks: T::BlockNumber) -> DispatchResult

Schedule a resume of GRANDPA after pausing.

pub fn schedule_change(
next_authorities: AuthorityList,
in_blocks: T::BlockNumber,
forced: Option<T::BlockNumber>
) -> DispatchResult

Schedule a change in the authorities.

The change will be applied at the end of execution of the block in_blocks after the current block. This value may be 0, in which case the change is applied at the end of the current block.

If the forced parameter is defined, this indicates that the current set has been synchronously determined to be offline and that after in_blocks the given change should be applied. The given block number indicates the median last finalized block number and it should be used as the canon block when starting the new grandpa voter.

No change should be signaled while any change is pending. Returns an error if a change is already pending.

pub fn submit_unsigned_equivocation_report(
equivocation_proof: EquivocationProof<T::Hash, T::BlockNumber>,
key_owner_proof: T::KeyOwnerProof
) -> Option<()>

Submits an extrinsic to report an equivocation. This method will create an unsigned extrinsic with a call to report_equivocation_unsigned and will push the transaction to the pool. Only useful in an offchain context.