Crate pallet_grandpa

GRANDPA Consensus module for runtime.

This manages the GRANDPA authority set ready for the native code. These authorities are only for GRANDPA finality, not for consensus overall.

In the future, it will also handle misbehavior reports, and on-chain finality notifications.

For full integration with GRANDPA, the GrandpaApi should be implemented. The necessary items are re-exported via the fg_primitives crate.

Re-exports

pub use sp_consensus_grandpa as fg_primitives;

pub use pallet::*;

Modules

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

EquivocationOffence

GRANDPA equivocation offence report.

EquivocationReportSystem

GRANDPA equivocation offence report system.

StoredPendingChange

A stored pending change. Limit is the bound for next_authorities

TimeSlot

A round number and set id which point on the time of an offence.

Enums

StoredState

Current state of the GRANDPA authority set. State transitions must happen in the same order of states defined below, e.g. Paused implies a prior PendingPause.

Traits

WeightInfo

Type Aliases

AuthorityId

Identity of a Grandpa authority.

AuthorityList

A list of Grandpa authorities with associated weights.

AuthorityWeight

The weight of an authority.

BoundedAuthorityList

Bounded version of AuthorityList, Limit being the bound