Struct pallet_grandpa::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

Methods for the ValidateUnsigned implementation: It restricts calls to report_equivocation_unsigned to local calls (i.e. extrinsics generated on this node) or that already in a block. This guarantees that only block authors can include unsigned equivocation reports.

pub fn validate_unsigned(
source: TransactionSource,
call: &Call<T>
) -> TransactionValidity

pub fn pre_dispatch(call: &Call<T>) -> Result<(), TransactionValidityError>

impl<T: Config> Pallet<T>

pub fn report_equivocation(
origin: OriginFor<T>,
equivocation_proof: EquivocationProof<T::Hash, T::BlockNumber>,
key_owner_proof: T::KeyOwnerProof
) -> DispatchResultWithPostInfo

Report voter equivocation/misbehavior. This method will verify the equivocation proof and validate the given key ownership proof against the extracted offender. If both are valid, the offence will be reported.

pub fn report_equivocation_unsigned(
origin: OriginFor<T>,
equivocation_proof: EquivocationProof<T::Hash, T::BlockNumber>,
key_owner_proof: T::KeyOwnerProof
) -> DispatchResultWithPostInfo

Report voter equivocation/misbehavior. This method will verify the equivocation proof and validate the given key ownership proof against the extracted offender. If both are valid, the offence will be reported.

This extrinsic must be called unsigned and it is expected that only block authors will call it (validated in ValidateUnsigned), as such if the block author is defined it will be defined as the equivocation reporter.

pub fn note_stalled(
origin: OriginFor<T>,
delay: T::BlockNumber,
best_finalized_block_number: T::BlockNumber
) -> DispatchResultWithPostInfo

Note that the current authority set of the GRANDPA finality gadget has stalled. This will trigger a forced authority set change at the beginning of the next session, to be enacted delay blocks after that. The delay should be high enough to safely assume that the block signalling the forced change will not be re-orged (e.g. 1000 blocks). The GRANDPA voters will start the new authority set using the given finalized block as base. Only callable by root.

impl<T: Config> Pallet<T>

pub fn state() -> StoredState<T::BlockNumber>

State of the current authority set.

impl<T: Config> Pallet<T>

pub fn pending_change() -> Option<StoredPendingChange<T::BlockNumber>>

Pending change: (signaled at, scheduled change).

impl<T: Config> Pallet<T>

pub fn next_forced() -> Option<T::BlockNumber>

next block number where we can force a change.

impl<T: Config> Pallet<T>

pub fn stalled() -> Option<(T::BlockNumber, T::BlockNumber)>

true if we are currently stalled.

impl<T: Config> Pallet<T>

pub fn current_set_id() -> SetId

The number of changes (both in terms of keys and underlying economic responsibilities) in the “set” of Grandpa validators from genesis.

impl<T: Config> Pallet<T>

pub fn session_for_set<KArg>(k: KArg) -> Option<SessionIndex> where
KArg: EncodeLike<SetId>,

A mapping from grandpa set ID to the index of the most recent session for which its members were responsible.

TWOX-NOTE: SetId is not under user control.

impl<T: Config> Pallet<T>

pub fn grandpa_authorities() -> AuthorityList

Get the current set of authorities, along with their respective weights.

pub fn schedule_pause(in_blocks: T::BlockNumber) -> DispatchResult

Schedule GRANDPA to pause starting in the given number of blocks. Cannot be done when already paused.

pub fn schedule_resume(in_blocks: T::BlockNumber) -> DispatchResult

Schedule a resume of GRANDPA after pausing.

pub fn schedule_change(
next_authorities: AuthorityList,
in_blocks: T::BlockNumber,
forced: Option<T::BlockNumber>
) -> DispatchResult

Schedule a change in the authorities.

The change will be applied at the end of execution of the block in_blocks after the current block. This value may be 0, in which case the change is applied at the end of the current block.

If the forced parameter is defined, this indicates that the current set has been synchronously determined to be offline and that after in_blocks the given change should be applied. The given block number indicates the median last finalized block number and it should be used as the canon block when starting the new grandpa voter.

No change should be signaled while any change is pending. Returns an error if a change is already pending.

pub fn submit_unsigned_equivocation_report(
equivocation_proof: EquivocationProof<T::Hash, T::BlockNumber>,
key_owner_proof: T::KeyOwnerProof
) -> Option<()>

Submits an extrinsic to report an equivocation. This method will create an unsigned extrinsic with a call to report_equivocation_unsigned and will push the transaction to the pool. Only useful in an offchain context.