Trait pallet_grandpa::GrandpaOffence

pub trait GrandpaOffence<FullIdentification>: Offence<FullIdentification> {
    fn new(
        session_index: SessionIndex,
        validator_set_count: u32,
        offender: FullIdentification,
        set_id: SetId,
        round: RoundNumber
    ) -> Self;
    fn new_time_slot(set_id: SetId, round: RoundNumber) -> Self::TimeSlot;
}

An interface for types that will be used as GRANDPA offences and must also implement the Offence trait. This trait provides a constructor that is provided all available data during processing of GRANDPA equivocations.

Required Methods

fn new(
session_index: SessionIndex,
validator_set_count: u32,
offender: FullIdentification,
set_id: SetId,
round: RoundNumber
) -> Self

Create a new GRANDPA offence using the given equivocation details.

fn new_time_slot(set_id: SetId, round: RoundNumber) -> Self::TimeSlot

Create a new GRANDPA offence time slot.